geo test

+++ geos-cluster/src/main/java/test/geos/matrix/MatrixParam.java
@@ -3,19 +3,19 @@ package test.geos.matrix;
 import java.io.File;
 
 public class MatrixParam {
-	private final int width;
-	private final int height;
+	private final long width;
+	private final long height;
 	private final File csv;
-	public MatrixParam(int width, int height, File csv) {
+	public MatrixParam(long width, long height, File csv) {
 		super();
 		this.width = width;
 		this.height = height;
 		this.csv = csv;
 	}
-	public int getWidth() {
+	public long getWidth() {
 		return width;
 	}
-	public int getHeight() {
+	public long getHeight() {
 		return height;
 	}
 	public File getCsv() {
